Change charge mode to JSON. Add support for timer charging
The XML based request was no longer working for my 2019 e-tron. Changed
the request to JSON.

Added a new "start_timed_charger" service call which switches the car
from "immediate" charging to "timer". The regular "start_charger"
service is able to switch from "timer" charging back to "immediate".

Note the "stop_charger" action does not work for my car - the request is
accepeted by the Audi server, queued for delivery to the car and
fetched, but the car appears to ignore the request and does not stop
charging.

Resolves #93

Signed-off-by: Matt Redfearn <matt.redfearn@gmail.com>